Separate hooks.ts into modules
