+++ src/app/features/editor/panels/DecksPanel.tsx
@@ -5,11 +5,12 @@ import { useMenu } from "../../../hooks/useMenu";
 import { Tree } from "../../../components/Tree";
 import { editorActions } from "../actions";
 import { selectors } from "../selectors";
-import { useConfirmDelete, usePromptCreate, usePromptRename } from "../hooks";
 import type { DeckId } from "../../../../api/services/game/types";
 import { PanelEmptyState } from "../components/PanelEmptyState";
 import { Panel } from "../components/Panel";
 import { ObjectIcon } from "../components/ObjectIcon";
+import { useConfirmDelete } from "../hooks/useConfirmDelete";
+import { usePromptCreate, usePromptRename } from "../hooks/usePromptCrud";
 import type { PanelProps } from "./definition";
 
 export function DecksPanel(props: PanelProps) {
